Create README.md
To bring in 100% transparency and to make the admissions highly cost-effective and to minimize the travel by the candidates and parents, The Student Counseling System is designed to enhance the counseling experience during the admission process. This system will provide great help not to the students but also to the parents who are there to invest money in children’s education.

The Student Counseling System is developing to improve the counseling, which fully works as online software. Through this system, the student will be able to report then register for the counseling as he is eligible after if the student is qualifying the criteria, then it will get the information regarding the colleges and the universities to which he can apply.

A student can also get information regarding how many seats have been booked and vacant seats available in the institution. This will help a student in selecting the branch in which they are interested and not what the colleges are offering. All this information will give in sort interval time to the student which speeds up the process reduce the institutes time.